+++ spec/apikey.go
@@ -26,7 +26,6 @@ import (
 	"crypto/sha256"
 	"encoding/hex"
 	"errors"
-	"reflect"
 	"sync"
 
 	"github.com/Sirupsen/logrus"
@@ -126,48 +125,6 @@ func (s *KeyFactory) Delete(obj interface{}) (interface{}, error) {
 	}
 	delete(s.keyMap, key.Spec.APIKeyData)
 	return actual, nil
-}
-
-// Contains reports whether the key store contains a particular key
-func (s *KeyFactory) Contains(params ...interface{}) (bool, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-	switch len(params) {
-	case 1:
-		switch p := params[0].(type) {
-		case string:
-			if _, ok := s.keyMap[p]; !ok {
-				return false, nil
-			}
-			return true, nil
-		case APIKey:
-			for _, v := range s.keyMap {
-				if reflect.DeepEqual(p, v) {
-					return true, nil
-				}
-			}
-			return false, nil
-		default:
-			return false, errors.New("could not recongized type of parameter")
-		}
-	case 2:
-		name, ok := params[0].(string)
-		if !ok {
-			return false, errors.New("first parameter should be a string")
-		}
-		namespace, ok := params[1].(string)
-		if !ok {
-			return false, errors.New("second parameter should be a string")
-		}
-		for _, v := range s.keyMap {
-			if v.ObjectMeta.Name == name && v.ObjectMeta.Namespace == namespace {
-				return true, nil
-			}
-		}
-		return false, nil
-	default:
-		return false, errors.New("too many parameters")
-	}
 }
 
 // IsEmpty reports whether the key store is empty