+++ spec/secret.go
@@ -119,28 +119,6 @@ func (s *SecretFactory) Delete(obj interface{}) (interface{}, error) {
 	return oldSecret, nil
 }
 
-// Contains reports whether the secrets store contains a particular secret
-func (s *SecretFactory) Contains(params ...interface{}) (bool, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-	if len(params) != 2 {
-		return false, errors.New("containers requires 2 params")
-	}
-	name, ok := params[0].(string)
-	if !ok {
-		return false, errors.New("first parameter should be a string")
-	}
-	namespace, ok := params[1].(string)
-	if !ok {
-		return false, errors.New("second parameter should be a string")
-	}
-	if _, ok := s.secretMap[namespace]; !ok {
-		return false, nil
-	}
-	_, ok = s.secretMap[namespace][name]
-	return ok, nil
-}
-
 // IsEmpty reports whether the secret store is empty
 func (s *SecretFactory) IsEmpty() bool {
 	s.mutex.RLock()